% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_within_between}
\alias{check_within_between}
\title{Test whether a predictor varies (or not) per participant}
\usage{
check_within_between(
  data,
  participant,
  predictor,
  tol = 0,
  min_within_prop = 0.9
)
}
\description{
Test whether a predictor varies (or not) per participant
}
\keyword{internal}
