% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eeg_data}
\alias{eeg_data}
\title{Simulated EEG data}
\format{
A data frame with 502000 rows and 5 variables:
\describe{
  \item{participant}{Character variable indicating the participant's ID.}
  \item{condition}{Integer variable indicating the condition.}
  \item{trial}{Integer variable indicating the trial number.}
  \item{time}{Numeric variable indicating the time step (in seconds).}
  \item{eeg}{Numeric variable indicating the EEG signal.}
}
}
\usage{
eeg_data
}
\description{
Simulated EEG data.
}
\keyword{datasets}
