% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_summary_data}
\alias{make_summary_data}
\title{Summarise trial-level data for \code{testing_through_time()}}
\usage{
make_summary_data(
  data,
  participant_id = "participant",
  outcome_id = "eeg",
  time_id = "time",
  predictor_id = NA,
  trials_id = NULL,
  family = gaussian(),
  multilevel = c("summary", "group"),
  na_rm = TRUE,
  continuous_agg = c("mean", "first")
)
}
\description{
Summarise trial-level data for \code{testing_through_time()}
}
\keyword{internal}
