% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_take_n_times}
\alias{st_take_n_times}
\title{Select approximately equally spaced times}
\usage{
st_take_n_times(time_vec, N)
}
\arguments{
\item{time_vec}{Numeric vector.}

\item{N}{Integer number of points.}
}
\value{
Subset of \code{time_vec}.
}
\description{
Select approximately equally spaced times
}
